airflow_setup/requirements.txt
boto3
python-dotenv
pandas-gbq